+++ client/components/FormattingBar/Controls/ControlsButton.js
