+++ client/components/FormattingBar/Controls/ControlsEquation.js
@@ -1,140 +1,103 @@
-import React, { Component } from 'react';
+/* eslint-disable react/no-danger */
+import React, { useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Button } from '@blueprintjs/core';
+import { useDebounce } from 'use-debounce';
+
 import { renderLatexString } from 'utils';
-import Icon from 'components/Icon/Icon';
+
+import { ControlsButton, ControlsButtonGroup } from './ControlsButton';
+
+require('./controls.scss');
 
 const propTypes = {
-	attrs: PropTypes.object.isRequired,
-	updateAttrs: PropTypes.func.isRequired,
-	changeNode: PropTypes.func.isRequired,
-	selectedNode: PropTypes.object.isRequired,
-	editorChangeObject: PropTypes.object.isRequired,
-	isSmall: PropTypes.bool.isRequired,
+	onClose: PropTypes.func.isRequired,
+	editorChangeObject: PropTypes.shape({
+		changeNode: PropTypes.func.isRequired,
+		updateNode: PropTypes.func.isRequired,
+		selectedNode: PropTypes.shape({
+			type: PropTypes.shape({
+				name: PropTypes.string,
+			}),
+			attrs: PropTypes.shape({
+				value: PropTypes.string.isRequired,
+				html: PropTypes.string,
+			}),
+		}),
+	}).isRequired,
 };
 
-class ControlsEquation extends Component {
-	constructor(props) {
-		super(props);
-		this.state = {
-			isEditing: false,
-			editingValue: '',
-		};
-		this.handleFocus = this.handleFocus.bind(this);
-		this.handleBlur = this.handleBlur.bind(this);
-		this.handleValueChange = this.handleValueChange.bind(this);
-		this.handleHTMLChange = this.handleHTMLChange.bind(this);
-		this.changeToInline = this.changeToInline.bind(this);
-		this.changeToBlock = this.changeToBlock.bind(this);
-	}
+const getSchemaDefinitionForNodeType = (editorChangeObject, nodeTypeName) => {
+	return editorChangeObject.view.state.schema.nodes[nodeTypeName];
+};
 
-	handleFocus() {
-		this.setState({
-			isEditing: true,
-			editingValue: this.props.attrs.value,
-		});
-	}
+const ControlsEquation = (props) => {
+	const { editorChangeObject, pendingAttrs, onClose } = props;
+	const { changeNode, selectedNode } = editorChangeObject;
+	const {
+		commitChanges,
+		hasPendingChanges,
+		updateAttrs,
+		attrs: { value, html },
+	} = pendingAttrs;
+	const [debouncedValue] = useDebounce(value, 250);
+	const hasMountedRef = useRef(false);
+	const isBlock = selectedNode.type.name === 'block_equation';
 
-	handleBlur() {
-		this.setState({
-			isEditing: false,
-		});
-	}
-
-	handleValueChange(evt) {
-		this.setState({ editingValue: evt.target.value });
-		const isBlock = this.props.selectedNode.type.name === 'block_equation';
-		renderLatexString(evt.target.value, isBlock, this.handleHTMLChange);
-	}
-
-	handleHTMLChange(html) {
-		this.props.updateAttrs({ value: this.state.editingValue, html: html });
-	}
-
-	changeToInline() {
-		const isBlock = this.props.selectedNode.type.name === 'block_equation';
-		if (isBlock) {
-			const nodeType = this.props.editorChangeObject.view.state.schema.nodes.equation;
-			renderLatexString(this.props.attrs.value, false, (newHtml) => {
-				this.props.changeNode(
-					nodeType,
-					{
-						value: this.props.attrs.value,
-						html: newHtml,
-					},
-					null,
-				);
-			});
+	useEffect(() => {
+		// Avoid an initial call to the server's LaTeX renderer on mount
+		// We shouldn't need this anyway -- but moreover, it will sometimes produce HTML that is
+		// insubstantially different from that given in our Prosemirror schema definitions, making
+		// it appear as though there is a user-driven change to the node that needs to be committed
+		// or reverted.
+		if (!hasMountedRef.current) {
+			hasMountedRef.current = true;
+			return;
 		}
-	}
+		renderLatexString(debouncedValue, false, (nextHtml) => {
+			updateAttrs({ html: nextHtml });
+		});
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [debouncedValue, isBlock]);
 
-	changeToBlock() {
-		const isBlock = this.props.selectedNode.type.name === 'block_equation';
-		if (!isBlock) {
-			const nodeType = this.props.editorChangeObject.view.state.schema.nodes.block_equation;
-			renderLatexString(this.props.attrs.value, true, (newHtml) => {
-				this.props.changeNode(
-					nodeType,
-					{
-						value: this.props.attrs.value,
-						html: newHtml,
-					},
-					null,
-				);
-			});
-		}
-	}
+	const handleUpdate = () => {
+		commitChanges();
+		onClose();
+	};
 
-	render() {
-		// <div>TODO: Not all of these are equations. Rename it to 'math'</div>
-		const isBlock = this.props.selectedNode.type.name === 'block_equation';
-		const iconSize = this.props.isSmall ? 12 : 16;
+	const handleChangeNodeType = () => {
+		const targetNodeType = isBlock ? 'equation' : 'block_equation';
+		const schemaDefinition = getSchemaDefinitionForNodeType(editorChangeObject, targetNodeType);
+		commitChanges();
+		changeNode(schemaDefinition, { value: value, html: html }, null);
+	};
 
-		return (
-			<div
-				className={`formatting-bar_controls-component ${this.props.isSmall ? 'small' : ''}`}
-			>
-				{/*  LaTex Adjustment */}
-				<div className="block">
-					<div className="label">LaTeX</div>
-					<div className="input wide">
-						<textarea
-							placeholder="Enter LaTeX math"
-							className="bp3-input bp3-fill"
-							value={
-								this.state.isEditing
-									? this.state.editingValue
-									: this.props.attrs.value
-							}
-							onFocus={this.handleFocus}
-							onBlur={this.handleBlur}
-							onChange={this.handleValueChange}
-						/>
-					</div>
-				</div>
-
-				{/*  Display Adjustment */}
-				<div className="block">
-					<div className="label">Display</div>
-					<div className="input">
-						<Button
-							onClick={this.changeToInline}
-							text="Inline"
-							icon={<Icon icon="align-left" iconSize={iconSize} />}
-							active={!isBlock}
-						/>
-						<Button
-							onClick={this.changeToBlock}
-							text="Block"
-							icon={<Icon icon="align-justify" iconSize={iconSize} />}
-							active={isBlock}
-						/>
-					</div>
-				</div>
+	return (
+		<div className="controls-equation-component">
+			<div className="section" key="structured">
+				<textarea
+					className="latex"
+					placeholder="Enter LaTeX"
+					value={value}
+					onChange={(evt) => updateAttrs({ value: evt.target.value })}
+				/>
 			</div>
-		);
-	}
-}
+			{html && (
+				<div className="section preview">
+					<div className="title">Preview</div>
+					<div className="preview" dangerouslySetInnerHTML={{ __html: html }} />
+					<ControlsButtonGroup>
+						<ControlsButton onClick={handleChangeNodeType}>
+							Change to {isBlock ? 'inline' : 'block'}
+						</ControlsButton>
+						<ControlsButton disabled={!hasPendingChanges} onClick={handleUpdate}>
+							Update equation
+						</ControlsButton>
+					</ControlsButtonGroup>
+				</div>
+			)}
+		</div>
+	);
+};
 
 ControlsEquation.propTypes = propTypes;
 export default ControlsEquation;
