+++ client/components/FormattingBar/Controls/ControlsFootnoteCitation.js
