+++ client/components/FormattingBar/Controls/ControlsLink.js
