+++ client/components/FormattingBar/Controls/ControlsMedia/AlignmentControl.js
