+++ client/components/FormattingBar/Controls/ControlsMedia/ControlsMedia.js
