+++ client/components/FormattingBar/Controls/ControlsMedia/SliderInputControl.js
