+++ client/components/FormattingBar/Controls/ControlsTable.js
@@ -1,131 +1,61 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { Menu, MenuItem, MenuDivider } from '@blueprintjs/core';
-import DropdownButton from 'components/DropdownButton/DropdownButton';
+import { Toolbar, ToolbarItem, useToolbarState } from 'reakit';
+import { Button } from '@blueprintjs/core';
 
-const propTypes = {
-	menuItems: PropTypes.array.isRequired,
-	isSmall: PropTypes.bool.isRequired,
-	editorChangeObject: PropTypes.object.isRequired,
-};
+import CommandMenu from '../CommandMenu';
+
+const rowCommands = [
+	{ key: 'table-add-row-before', title: 'Add row before', icon: 'add-row-top' },
+	{ key: 'table-add-row-after', title: 'Add row after', icon: 'add-row-bottom' },
+	{ key: 'table-toggle-header-row', title: 'Toggle header row', icon: 'page-layout' },
+	{ key: 'table-delete-row', title: 'Delete row', icon: 'disable' },
+];
+
+const columnCommands = [
+	{ key: 'table-add-column-before', title: 'Add column before', icon: 'add-column-left' },
+	{ key: 'table-add-column-after', title: 'Add column after', icon: 'add-column-right' },
+	{ key: 'table-toggle-header-column', title: 'Toggle column row', icon: 'page-layout' },
+	{ key: 'table-delete-column', title: 'Delete column', icon: 'disable' },
+];
+
+const buttonCommands = [
+	{ key: 'table-merge-cells', title: 'Merge cells', icon: 'merge-columns' },
+	{ key: 'table-split-cell', title: 'Split cells', icon: 'split-columns' },
+	{ key: 'toggle-header-cell', title: 'Toggle header cells', icon: 'header' },
+	{ key: 'table-delete', title: 'Remove table', icon: 'trash' },
+];
 
 const ControlsTable = (props) => {
-	const commands = {};
-	props.menuItems.forEach((menuItem) => {
-		commands[menuItem.title] = menuItem;
-	});
+	const { editorChangeObject } = props;
+	const { menuItems = [], view } = editorChangeObject;
+	const toolbar = useToolbarState({ loop: true });
+
+	// eslint-disable-next-line react/prop-types
+	const renderDisclosure = ({ ref, ...disclosureProps }) => {
+		return (
+			<Button
+				minimal
+				className="block-type-selector-component"
+				rightIcon="caret-down"
+				elementRef={ref}
+				icon="th"
+				{...disclosureProps}
+			/>
+		);
+	};
 
 	return (
-		<div className={`formatting-bar_controls-component ${props.isSmall ? 'small' : ''}`}>
-			{!props.isSmall && <div className="separator" />}
-
-			<DropdownButton label="Table" isMinimal={true} usePortal={false}>
-				<Menu>
-					<MenuDivider title="Rows" />
-					<MenuItem
-						text="Add Row Before"
-						disabled={!commands['table-add-row-before'].isActive}
-						onClick={() => {
-							commands['table-add-row-before'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Add Row After"
-						disabled={!commands['table-add-row-after'].isActive}
-						onClick={() => {
-							commands['table-add-row-after'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Toggle Header Row"
-						disabled={!commands['table-toggle-header-row'].isActive}
-						onClick={() => {
-							commands['table-toggle-header-row'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Remove Row"
-						disabled={!commands['table-delete-row'].isActive}
-						onClick={() => {
-							commands['table-delete-row'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuDivider title="Columns" />
-					<MenuItem
-						text="Add Column Before"
-						disabled={!commands['table-add-column-before'].isActive}
-						onClick={() => {
-							commands['table-add-column-before'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Add Column After"
-						disabled={!commands['table-add-column-after'].isActive}
-						onClick={() => {
-							commands['table-add-column-after'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Toggle Header Column"
-						disabled={!commands['table-toggle-header-column'].isActive}
-						onClick={() => {
-							commands['table-toggle-header-column'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Remove Column"
-						disabled={!commands['table-delete-column'].isActive}
-						onClick={() => {
-							commands['table-delete-column'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuDivider title="Cells" />
-					<MenuItem
-						text="Merge Cells"
-						disabled={!commands['table-merge-cells'].isActive}
-						onClick={() => {
-							commands['table-merge-cells'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Split Cell"
-						disabled={!commands['table-split-cell'].isActive}
-						onClick={() => {
-							commands['table-split-cell'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuItem
-						text="Toggle Header Cell"
-						disabled={!commands['table-toggle-header-cell'].isActive}
-						onClick={() => {
-							commands['table-toggle-header-cell'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-					<MenuDivider />
-					<MenuItem
-						text="Remove Table"
-						disabled={!commands['table-delete'].isActive}
-						onClick={() => {
-							commands['table-delete'].run();
-							props.editorChangeObject.view.focus();
-						}}
-					/>
-				</Menu>
-			</DropdownButton>
-		</div>
+		<Toolbar {...toolbar} className="controls-table-component" aria-label="Table options">
+			<ToolbarItem
+				aria-label="Table options"
+				as={CommandMenu}
+				disclosure={renderDisclosure}
+				commands={[rowCommands, columnCommands, buttonCommands]}
+				editorChangeObject={editorChangeObject}
+				{...toolbar}
+			/>
+		</Toolbar>
 	);
 };
 
-ControlsTable.propTypes = propTypes;
 export default ControlsTable;
