+++ client/components/FormattingBar/FormattingBarButton.js
