+++ client/components/FormattingBar/FormattingBarMediaButton.js
