+++ client/components/FormattingBar/FormattingBarPopover.js
