+++ client/components/FormattingBar/Media/Media.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Menu, MenuItem } from '@blueprintjs/core';
+
 import { Icon } from 'components';
+
 import MediaImage from './MediaImage';
 import MediaVideo from './MediaVideo';
 import MediaAudio from './MediaAudio';
@@ -48,20 +50,6 @@ class Media extends Component {
 			{ text: 'Codepen', icon: 'codepen' },
 			{ text: 'Vimeo', icon: 'vimeo' },
 			{ text: 'SoundCloud', icon: 'soundcloud' },
-
-			// TODO: Not yet implemented
-			// { text: 'Spotify', icon: 'spotify' },
-			// { text: 'Twitter', icon: 'twitter' },
-			// { text: 'Github', icon: 'github' },
-			// { text: 'Prezi', icon: 'prezi' },
-			// { text: 'LinkedIn', icon: 'linkedin' },
-			// { text: 'Giphy', icon: 'giphy' },
-			// { text: 'Unsplash', icon: 'unsplash' },
-			// { text: 'Google Maps', icon: 'googlemaps' },
-			// { text: 'Slideshare', icon: 'slideshare' },
-
-			// { text: 'Facebook', icon: 'facebook' },
-			// { text: 'Reddit', icon: 'reddit' }, /* oEmbed URL is not functioning as of 12/2018 */
 		];
 		const activeItem = this.state.activeItem;
 		const componentProps = {
