+++ client/components/FormattingBar/buttons.js
