+++ client/components/FormattingBar/index.js
