+++ client/components/FormattingBar/positioning.js
