+++ client/components/FormattingBarLegacy/Controls/Controls.js
