+++ client/components/FormattingBarLegacy/Controls/ControlsAudio.js
