+++ client/components/FormattingBarLegacy/Controls/ControlsFootnote.js
