+++ client/components/FormattingBarLegacy/Controls/ControlsIframe.js
