+++ client/components/FormattingBarLegacy/Controls/ControlsImage.js
