+++ client/components/FormattingBarLegacy/Controls/controls.scss
@@ -1,0 +1,68 @@
+.formatting-bar_controls-component {
+	display: flex;
+	flex-wrap: wrap;
+	.block {
+		margin-right: 1em;
+		.label {
+			margin-bottom: 2px;
+			font-size: 14px;
+			&.over-buttons {
+				position: relative;
+				left: 7px;
+			}
+		}
+		.input {
+			display: flex;
+			align-items: center;
+			min-height: 26px;
+			&.wide {
+				width: 200px;
+				word-break: break-word;
+			}
+			.bp3-numeric-input {
+				input {
+					width: 150px;
+					height: 20px;
+				}
+				.input-suffix {
+					padding: 1px 10px;
+					opacity: 0.7;
+				}
+			}
+		}
+	}
+	.simple-editor-wrapper {
+		border-radius: 3px;
+		box-shadow: 0 0 0 0 rgba(19, 124, 189, 0), 0 0 0 0 rgba(19, 124, 189, 0),
+			inset 0 0 0 1px rgba(16, 22, 26, 0.15), inset 0 1px 1px rgba(16, 22, 26, 0.2);
+		padding: 0.5em;
+		background: white;
+		width: 100%;
+		// height: 50px;
+		height: 45px;
+		overflow-y: scroll;
+	}
+	textarea {
+		height: 45px;
+		min-height: 45px;
+	}
+	.file-select {
+		input.file-input {
+			display: none;
+		}
+	}
+	&.small {
+		.block .label,
+		.hide-on-small {
+			display: none;
+		}
+		.simple-editor-wrapper {
+			font-size: 12px;
+			height: 22px;
+			padding: 0.3em;
+		}
+		.input {
+			min-height: 22px;
+		}
+	}
+}
