+++ client/components/FormattingBarLegacy/Controls/discussionAutocomplete.scss
