+++ client/components/FormattingBarLegacy/Controls/index.js
