+++ client/components/FormattingBarLegacy/Media/MediaImage.js
