+++ client/components/FormattingBarLegacy/Media/index.js
