+++ client/components/FormattingBarLegacy/Media/media.scss
