+++ client/components/FormattingBarLegacy/formattingBar.scss
@@ -1,0 +1,68 @@
+.formatting-bar-component {
+	display: flex;
+	max-width: 100%;
+	.bp3-button {
+		min-width: auto;
+		min-height: auto;
+		padding: 5px 14px;
+		margin: 0em 1px;
+	}
+	.dropdown-button {
+		padding: 5px;
+		min-height: auto;
+		white-space: nowrap;
+	}
+	.separator {
+		display: inline-block;
+		width: 1px;
+		height: 26px;
+		vertical-align: middle;
+		background: #ccc;
+		margin: 0px 3px;
+	}
+	.link-formatting {
+		display: inline-block;
+		padding-left: 5px;
+	}
+	.short-title {
+		display: none;
+	}
+	&.small {
+		.bp3-button {
+			padding: 5px 10px;
+		}
+		.bp3-button-text {
+			line-height: 12px;
+		}
+		.dropdown-button {
+			padding: 3px;
+		}
+		.separator {
+			height: 22px;
+		}
+		.link-formatting {
+			padding-left: 2px;
+			.bp3-input {
+				height: 18px;
+				line-height: 18px;
+			}
+		}
+		.short-title {
+			display: inline;
+		}
+		.full-title {
+			display: none;
+		}
+	}
+}
+
+@media only screen and (max-width: 750px) {
+	.formatting-bar-component {
+		.short-title {
+			display: inline;
+		}
+		.full-title {
+			display: none;
+		}
+	}
+}
